fix(extrinsics): remove output_json parameter from extrinsics structs
