+++ crates/cargo-contract/Cargo.toml
@@ -30,6 +30,7 @@ which = "4.4.0"
 colored = "2.0.4"
 serde_json = "1.0.105"
 url = { version = "2.4.1", features = ["serde"] }
+serde = { version = "1.0.183", default-features = false, features = ["derive"] }
 
 
 # dependencies for extrinsics (deploying and calling a contract)
