+++ crates/cargo-contract/src/cmd/upload.rs
@@ -6,11 +6,10 @@ use anyhow::Result;
 use contract_build::name_value_println;
 use contract_extrinsics::{
     display_dry_run_result_warning,
-    CodeHashResult,
+    Balance,
     ExtrinsicOpts,
     TokenMetadata,
     UploadCommandBuilder,
-    UploadDryRunResult,
 };
 
 #[derive(Debug, clap::Args)]
@@ -34,7 +33,6 @@ pub fn handle_upload(upload_command: &UploadCommand) -> Result<(), ErrorVariant>
     Runtime::new()?.block_on(async {
         let upload_exec = UploadCommandBuilder::default()
             .extrinsic_opts(upload_command.extrinsic_opts.clone())
-            .output_json(upload_command.output_json())
             .done()
             .await;
 
@@ -48,7 +46,7 @@ pub fn handle_upload(upload_command: &UploadCommand) -> Result<(), ErrorVariant>
                         code_hash: format!("{:?}", result.code_hash),
                         deposit: result.deposit,
                     };
-                    if upload_exec.output_json() {
+                    if upload_command.output_json() {
                         println!("{}", upload_result.to_json()?);
                     } else {
                         upload_result.print();
@@ -58,7 +56,7 @@ pub fn handle_upload(upload_command: &UploadCommand) -> Result<(), ErrorVariant>
                 Err(err) => {
                     let metadata = upload_exec.client().metadata();
                     let err = ErrorVariant::from_dispatch_error(&err, &metadata)?;
-                    if upload_exec.output_json() {
+                    if upload_command.output_json() {
                         return Err(err)
                     } else {
                         name_value_println!("Result", err);
@@ -68,7 +66,7 @@ pub fn handle_upload(upload_command: &UploadCommand) -> Result<(), ErrorVariant>
         } else {
             let upload_result = upload_exec.upload_code().await?;
             let display_events = upload_result.display_events;
-            let output = if upload_exec.output_json() {
+            let output = if upload_command.output_json() {
                 display_events.to_json()?
             } else {
                 let token_metadata = TokenMetadata::query(upload_exec.client()).await?;
@@ -80,7 +78,7 @@ pub fn handle_upload(upload_command: &UploadCommand) -> Result<(), ErrorVariant>
                 let upload_result = CodeHashResult {
                     code_hash: format!("{:?}", code_stored.code_hash),
                 };
-                if upload_exec.output_json() {
+                if upload_command.output_json() {
                     println!("{}", upload_result.to_json()?);
                 } else {
                     upload_result.print();
@@ -95,4 +93,38 @@ pub fn handle_upload(upload_command: &UploadCommand) -> Result<(), ErrorVariant>
         }
         Ok(())
     })
+}
+
+#[derive(serde::Serialize)]
+pub struct CodeHashResult {
+    pub code_hash: String,
+}
+
+impl CodeHashResult {
+    pub fn to_json(&self) -> Result<String> {
+        Ok(serde_json::to_string_pretty(self)?)
+    }
+
+    pub fn print(&self) {
+        name_value_println!("Code hash", format!("{:?}", self.code_hash));
+    }
+}
+
+#[derive(serde::Serialize)]
+pub struct UploadDryRunResult {
+    pub result: String,
+    pub code_hash: String,
+    pub deposit: Balance,
+}
+
+impl UploadDryRunResult {
+    pub fn to_json(&self) -> Result<String> {
+        Ok(serde_json::to_string_pretty(self)?)
+    }
+
+    pub fn print(&self) {
+        name_value_println!("Result", self.result);
+        name_value_println!("Code hash", format!("{:?}", self.code_hash));
+        name_value_println!("Deposit", format!("{:?}", self.deposit));
+    }
 }
